added text input simple class
added text input class, which can handle user input with customized options.
this is currently not, however, possible the UTF8 characters, because engine may have slight issues looping UTF8 strings.
this is also currently not possible to speak capital letters, as the function to check whether the character is a capital letter is missing.
none the less, a simple input is now available.